# docs/swagger.yaml
definitions:
  pod.Container:
    properties:
      container_name:
        type: string
      namespace:
        type: string
      pod_name:
        type: string
      usages:
        additionalProperties:
          $ref: '#/definitions/resource.ResourceUsageInfo'
        description: Resource usage list
        type: object
    type: object
  pod.Pod:
    properties:
      containers:
        description: Container information
        items:
          $ref: '#/definitions/pod.Container'
        type: array
      name:
        type: string
      namespace:
        type: string
      usage:
        additionalProperties:
          $ref: '#/definitions/resource.ResourceUsageInfo'
        description: total usage infromation
        type: object
    type: object
  resource.ResourceUsageInfo:
    properties:
      current_usage:
        type: number
      limit:
        type: number
      name:
        type: string
      optimized_usage:
        type: number
      request:
        type: number
      status:
        type: string
      usage:
        items:
          $ref: '#/definitions/resource.TimeSeriesDatapoint'
        type: array
    type: object
  resource.TimeSeriesDatapoint:
    properties:
      time:
        type: integer
      value:
        type: number
    type: object
  vm.Vm:
    properties:
      name:
        type: string
      usages:
        additionalProperties:
          $ref: '#/definitions/resource.ResourceUsageInfo'
        description: Resource usage list
        type: object
    type: object
info:
  contact: {}
paths:
  /api/v1/pods:
    get:
      consumes:
      - application/json
      description: pod의 리소스 quota 정보와 사용량 및 사용량 기반의 최적 사용량을 제공한다.
      parameters:
      - description: the name of pod
        in: path
        name: name
        type: string
      - description: the namespace of pod
        in: path
        name: namespace
        type: string
      - description: start time
        in: query
        name: start
        type: string
      - description: end time
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pod.Pod'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: pod의 리소스 정보 및 사용량 관련 정보 제공
  /api/v1/pods/{namespace}/{name}/forecast:
    get:
      consumes:
      - application/json
      description: Create forecast task and result task UUID
      parameters:
      - description: the namespace of pod
        in: path
        name: namespace
        required: true
        type: string
      - description: the name of pod
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Post pod forecast task
  /api/v1/pods/{uuid}/forecast/result:
    get:
      consumes:
      - application/json
      description: Get forecast result
      parameters:
      - description: the uuid of forecast task
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get vm forecast result
  /api/v1/pods/forecast:
    get:
      consumes:
      - application/json
      description: forecast 작업이 끝나지 않은 경우 nil 값 제공
      parameters:
      - description: the name of pod
        in: path
        name: name
        required: true
        type: string
      - description: the namespace of pod
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: 특정 pod의 forecast 결과 제공
  /api/v1/pods/forecast/{uuid}/result:
    get:
      consumes:
      - application/json
      description: id를 통해서 forecast 결과를 제공함. 만약 작업이 끝나지 않은 경우 nil 값 제공.
      parameters:
      - description: the uuid of forecast task
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: 사용자의 요청에 따라 발급한 forecast id를 통해서 forecast 결과를 제공
  /api/v1/pods/forecast/{uuid}/status:
    get:
      consumes:
      - application/json
      parameters:
      - description: the uuid of forecast task
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: 사용자의 요청에 따라 발급한 forecast id를 통해서 forecast 완료 여부 제공
  /api/v1/pods/forecast/status:
    get:
      consumes:
      - application/json
      parameters:
      - description: the name of pod
        in: path
        name: name
        required: true
        type: string
      - description: the namespace of pod
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: 특정 pod의 forecast 완료 여부를 알려줌
  /api/v1/pods/resource-quota:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: 클러스터 전반적인 지표들을 제공
  /api/v1/vms/{name}:
    get:
      consumes:
      - application/json
      description: Get all resource usage history and optimization usage value
      parameters:
      - description: name of the vm
        in: path
        name: name
        required: true
        type: string
      - description: start time
        in: query
        name: start
        type: string
      - description: end time
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get vm usage history and optimization usage
  /api/v1/vms/{name}/forecast:
    get:
      consumes:
      - application/json
      description: Get forecast result from TimescaleDB
      parameters:
      - description: the name of vm
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get vm forecast result
  /api/v1/vms/{name}/forecast/status:
    get:
      consumes:
      - application/json
      description: Get forecast task from TimescaleDB
      parameters:
      - description: the name of vm
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get vm forecast task status
  /api/v1/vms/{name}/resource-quota:
    get:
      consumes:
      - application/json
      description: Get vm resource quota information from TimescaleDB
      parameters:
      - description: the name of vm
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get vm resource quota
  /api/v1/vms/{uuid}/forecast/status:
    get:
      consumes:
      - application/json
      description: Get forecast task status by UUID
      parameters:
      - description: the uuid of forecast task
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get vm forecast task status by UUID
  /api/v1/vms/resource-quota:
    get:
      consumes:
      - application/json
      description: Get all vm resource quota information from TimescaleDB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vm.Vm'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get all vm resource quota
swagger: "2.0"